% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{abc_smc_2}
\alias{abc_smc_2}
\title{simulate a tree using environmental diversification}
\usage{
abc_smc_2(
  m1,
  m2,
  m3,
  m_weights,
  max_w,
  batch_size,
  crown_age,
  min_lin,
  max_lin,
  num_threads,
  sd_p,
  self_prob_m
)
}
\arguments{
\item{crown_age}{crown age of the tree to be simulated}

\item{min_lin}{minimum number of lineages}

\item{max_lin}{maximum number of lineages}

\item{num_threads}{number of threads}

\item{model}{a vector of the four paramaters of the model}

\item{num_repl}{a vector that indicates the time points of water level changes}
}
\value{
newick string
}
\description{
simulate a tree using environmental diversification
}
