% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accept_from_R.R
\name{accept_from_R}
\alias{accept_from_R}
\title{calculate summary statistics for a focal tree. The following summary
  statistics are calculated:
  nLTT, gamma, mean branching times and number of lineages}
\usage{
accept_from_R(emp_stats, newick, threshold, sd, emp_brts)
}
\arguments{
\item{emp_stats}{statistics to compare against}

\item{newick}{vector with newick strings}

\item{threshold}{maximum threshold value}

\item{sd}{recorded standard deviation in the population}
}
\value{
vector with true or false for each tree
}
\description{
calculate summary statistics for a focal tree. The following summary
  statistics are calculated:
  nLTT, gamma, mean branching times and number of lineages
}
