% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fit.R
\name{calc_fit}
\alias{calc_fit}
\title{calculate the overall fit of the particle to the data}
\usage{
calc_fit(sum_stats, emp_stats)
}
\arguments{
\item{sum_stats}{summary statistics of the tree generated by the particle}

\item{emp_stats}{summary statistics of the empirical tree}
}
\value{
sum of normalized differences between the summary statistics
}
\description{
calculate the overall fit of the particle to the data
}
