% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accept_from_R.R
\name{calc_stat}
\alias{calc_stat}
\title{calculate statistics}
\usage{
calc_stat(focal_tree, index, brts_emp_tree)
}
\arguments{
\item{focal_tree}{focal tree}

\item{index}{index}

\item{brts_emp_tree}{branching times emperical tree}
}
\description{
calculate statistics
}
