% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sum_stats.R
\name{calc_sum_stats}
\alias{calc_sum_stats}
\title{calculate summary statistics for a focal tree. The following summary
  statistics are calculated:
  nLTT, gamma, mean branching times and number of lineages}
\usage{
calc_sum_stats(focal_tree)
}
\arguments{
\item{focal_tree}{input phylogenetic tree for which to calculate
summary statistics.}
}
\value{
vector of 70 summary statistics
}
\description{
calculate summary statistics for a focal tree. The following summary
  statistics are calculated:
  nLTT, gamma, mean branching times and number of lineages
}
