% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_weights}
\alias{calc_weights}
\title{calculate weights}
\usage{
calc_weights(m1, m2, m3, p, m_weights, sd_p, self_prob_m)
}
\arguments{
\item{m1}{matrix with particles from model 1}

\item{m2}{matrix with particles from model 2}

\item{m3}{matrix with particles from model 3}

\item{p}{matrix with particles for which to calculate weight}

\item{m_weights}{vector with model weights of m1, m2 and m3}

\item{sd_p}{standard deviation of parameter change}

\item{self_prob_m}{probability of drawing model}
}
\value{
vector with weights
}
\description{
calculate weights
}
