% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_weights_R.R
\name{calc_weights_R}
\alias{calc_weights_R}
\title{calculate weights}
\usage{
calc_weights_R(m1, m2, m3, p, m_w, sd_p, self_prob, num_threads = -1)
}
\arguments{
\item{m1}{previous particles of model 1}

\item{m2}{previous particles of model 2}

\item{m3}{previous particles of model 3}

\item{p}{current particles}

\item{m_w}{previous model weights}

\item{sd_p}{standard deviation of parameter perturbation}

\item{self_prob}{probability of drawing self model}

\item{num_threads}{number of threads}
}
\value{
vector of weights
}
\description{
calculate weights
}
