% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_envidiv_tree.R
\name{correct_ltable_internal}
\alias{correct_ltable_internal}
\title{correct ltable}
\usage{
correct_ltable_internal(input_matrix, crown_age)
}
\arguments{
\item{input_matrix}{ltable}

\item{crown_age}{age}
}
\description{
correct ltable
}
\keyword{export}
