% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{create_ref_table_cpp_serial}
\alias{create_ref_table_cpp_serial}
\title{simulate a tree using environmental diversification}
\usage{
create_ref_table_cpp_serial(
  model,
  num_repl,
  crown_age,
  min_lin,
  max_lin,
  num_threads
)
}
\arguments{
\item{model}{a vector of the four paramaters of the model}

\item{num_repl}{a vector that indicates the time points of water level changes}

\item{crown_age}{crown age of the tree to be simulated}

\item{min_lin}{minimum number of lineages}

\item{max_lin}{maximum number of lineages}

\item{num_threads}{number of threads}
}
\value{
newick string
}
\description{
simulate a tree using environmental diversification
}
