% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_reference_table.R
\name{create_reference_table}
\alias{create_reference_table}
\title{function to perform ABC-SMC}
\usage{
create_reference_table(simulations_per_model = 1000, min_tips = 50,
  max_tips = 150, model = NULL, emp_tree = NULL, crown_age = NULL,
  write_to_file = FALSE, file_name)
}
\arguments{
\item{min_tips}{minimum number of tips}

\item{max_tips}{maximum number of tips}

\item{model}{used water model}

\item{emp_tree}{phy object holding phylogeny of the tree to be fitted on}

\item{write_to_file}{boolean, if TRUE, results are written to file.}

\item{file_name}{file name}

\item{number_of_particles}{number of particles used per iteration of
the SMC algorithm}
}
\value{
a tibble containing the results
}
\description{
function to perform ABC-SMC
}
