% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_statistics_list.R
\name{create_statistics_list}
\alias{create_statistics_list}
\title{function wrapper around treestats}
\usage{
create_statistics_list()
}
\value{
list of statistics functions
}
\description{
function wrapper around treestats
}
