% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{generate_expon}
\alias{generate_expon}
\title{test exponential generator}
\usage{
generate_expon(lambda)
}
\arguments{
\item{lambda}{input parameter}
}
\value{
number drawn from exponential distribution
}
\description{
test exponential generator
}
