% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_from_prior.R
\name{generate_from_prior}
\alias{generate_from_prior}
\title{function to perform ABC-SMC}
\usage{
generate_from_prior(
  number_of_particles = 1000,
  min_tips = 50,
  max_tips = 150,
  model = NULL,
  emp_tree = NULL,
  crown_age = NULL,
  write_to_file = FALSE,
  file_name,
  exp_prior = FALSE
)
}
\arguments{
\item{number_of_particles}{number of particles used per iteration of
the SMC algorithm}

\item{min_tips}{minimum number of tips}

\item{max_tips}{maximum number of tips}

\item{model}{used water model}

\item{emp_tree}{phy object holding phylogeny of the tree to be fitted on}

\item{crown_age}{crown age}

\item{write_to_file}{boolean, if TRUE, results are written to file.}

\item{file_name}{file name}

\item{exp_prior}{use an exponential prior or not.}
}
\value{
a tibble containing the results
}
\description{
function to perform ABC-SMC
}
