% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_stack_simple.R
\name{generate_stack_simple}
\alias{generate_stack_simple}
\title{function to perform ABC-SMC}
\usage{
generate_stack_simple(
  number_of_replicates = 1000,
  focal_model = 1,
  min_tips = 50,
  max_tips = 150,
  emp_tree = NULL,
  crown_age = NULL,
  write_to_file = FALSE,
  file_name = NULL,
  num_threads = 1
)
}
\arguments{
\item{number_of_replicates}{number of particles used per iteration of
the SMC algorithm}

\item{min_tips}{minimum number of tips}

\item{max_tips}{maximum number of tips}

\item{emp_tree}{phy object holding phylogeny of the tree to be fitted on}

\item{crown_age}{crown age}

\item{write_to_file}{boolean, if TRUE, results are written to file.}

\item{file_name}{file name}

\item{num_threads}{number of threads}

\item{parameters}{parameters, a vector with:
1) extinction rate,
2) (sympatric) speciation rate at high water,
3) sympatric speciation rate at low water,
4) allopatric speciation rate at low water,
5) posterior perturbation and
6) the specific model, where 1) model without water level changes,
2) literature water level changes, 3) extrapolated water level changes,
4) standard birth-death model.}
}
\value{
a tibble containing the results
}
\description{
function to perform ABC-SMC
}
