% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions.R
\name{generate_water}
\alias{generate_water}
\title{generate a sequence of waterlevel changes, depending on the model}
\usage{
generate_water(water_model, maximum_time)
}
\arguments{
\item{water_model}{can be 1) no water level changes, 2) literature water level changes 3) literature water level changes + exponentially distributed water level changes}

\item{maximum_time}{crown age}
}
\value{
vector of water level changes in [0, maximum_time], where 0 indicates the start of the tree (e.g. the root)
}
\description{
generate a sequence of waterlevel changes, depending on the model
}
