% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_params_abcrf.R
\name{infer_params_abcrf}
\alias{infer_params_abcrf}
\title{infer parameter values}
\usage{
infer_params_abcrf(reference_table, focal_model, to_analyze)
}
\arguments{
\item{reference_table}{reference table}

\item{focal_model}{focal model}

\item{to_analyze}{matrix with summary statistics of tree to infer paramaters for}
}
\value{
tibble with parameter estimates
}
\description{
infer parameter values
}
