% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{initial_draw_from_prior}
\alias{initial_draw_from_prior}
\title{generate particles from the prior}
\usage{
initial_draw_from_prior(num_particles, crown_age, min_lin, max_lin, verbose)
}
\arguments{
\item{num_particles}{number of particles}

\item{crown_age}{crown age}

\item{min_lin}{minimum number of lineages}

\item{max_lin}{maximum number of lineages}

\item{verbose}{verbose output}
}
\value{
numeric matrix
}
\description{
generate particles from the prior
}
