% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_within_prior.R
\name{is_within_prior}
\alias{is_within_prior}
\title{function to determine if the parameter combination is within the prior}
\usage{
is_within_prior(params)
}
\arguments{
\item{params}{vector of parameters}
}
\value{
true or false
}
\description{
function to determine if the parameter combination is within the prior
}
