% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ltab_to_ltable_to_phylo}
\alias{ltab_to_ltable_to_phylo}
\title{internal ltab to ltable to phylo}
\usage{
ltab_to_ltable_to_phylo(ltab, crown_age, dropextinct = T)
}
\arguments{
\item{ltab}{ltab}

\item{crown_age}{crown age}

\item{dropextinct}{drop extinct}
}
\value{
newick
}
\description{
internal ltab to ltable to phylo
}
