% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ltable_to_phylo}
\alias{ltable_to_phylo}
\title{export from DDD package, returns newick instead of tree}
\usage{
ltable_to_phylo(L, dropextinct = T)
}
\arguments{
\item{L}{ltable}

\item{dropextinct}{drop extinct}
}
\value{
newick
}
\description{
export from DDD package, returns newick instead of tree
}
