% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc_functions.R
\name{mutate_param}
\alias{mutate_param}
\title{perturbate all parameters}
\usage{
mutate_param(params, local_sd)
}
\arguments{
\item{params}{input vector containing unperturbed parameters}

\item{local_sd}{standard deviation of the perturbation kernel}
}
\value{
perturbed parameters
}
\description{
perturbate all parameters
}
