% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_params.R
\name{mutate_params}
\alias{mutate_params}
\title{perturbate all parameters}
\usage{
mutate_params(params, local_sd)
}
\arguments{
\item{params}{input vector containing unperturbed parameters}

\item{local_sd}{standard deviation of the perturbation kernel}
}
\value{
perturbed parameters
}
\description{
perturbate all parameters
}
