% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_statistics_list.R
\name{names_statistics_list}
\alias{names_statistics_list}
\title{return a vector of statistics names used}
\usage{
names_statistics_list()
}
\value{
vector of names
}
\description{
return a vector of statistics names used
}
