% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{param_from_prior_model_cpp}
\alias{param_from_prior_model_cpp}
\title{draw parameter combinations from the prior}
\usage{
param_from_prior_model_cpp(model)
}
\arguments{
\item{model}{chosen model}
}
\value{
vector with 6 entries: extinction, sympatric speciation at high
water, sympatric speciation at low water, allopatric speciation, amount of
perturbation, the chosen water model
}
\description{
draw parameter combinations from the prior
}
