% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_abc_smc.R
\name{perform_abc_smc}
\alias{perform_abc_smc}
\title{perform abc smc}
\usage{
perform_abc_smc(
  emp_tree,
  num_particles = 1000,
  num_threads = -1,
  sd_p = 0.05,
  sd_self = 0.7,
  min_tips = 40,
  max_tips = 120,
  write_to_file = TRUE
)
}
\arguments{
\item{emp_tree}{phylogenetic tree to fit on}

\item{num_particles}{number of particles}

\item{num_threads}{number of threads}

\item{sd_p}{standard deviation of parameter perturbation}

\item{min_tips}{minimum number of tips}

\item{max_tips}{maximum number of tips}

\item{write_to_file}{write to file}

\item{self_prob}{probability of drawing self model}
}
\value{
vector of weights
}
\description{
perform abc smc
}
