% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{raw_tree}
\alias{raw_tree}
\title{Function to create a newick string}
\usage{
raw_tree(params, water_changes, crown_age, seed = NULL)
}
\arguments{
\item{params}{parameter vector}

\item{water_changes}{vector with parameter changes}

\item{crown_age}{crown age of tree to be simulated}

\item{seed}{pseudo random number generator seed}
}
\value{
newick string
}
\description{
Function to create a newick string
}
