% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sim_envidiv_cpp}
\alias{sim_envidiv_cpp}
\title{simulate a tree using environmental diversification}
\usage{
sim_envidiv_cpp(model, parameters, crown_age, max_lin)
}
\arguments{
\item{model}{chosen model}

\item{parameters}{a vector of parameters}

\item{crown_age}{crown age of the tree to be simulated}

\item{max_lin}{maximum number of lineages}

\item{min_lin}{minimum number of lineages}
}
\value{
newick string
}
\description{
simulate a tree using environmental diversification
}
