% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_new.R
\name{sim_envidiv_tree_new_cond}
\alias{sim_envidiv_tree_new_cond}
\title{simulate a tree using the environmental diversification model}
\usage{
sim_envidiv_tree_new_cond(
  model,
  crown_age,
  min_lin = 4,
  max_lin = 500,
  num_tries = 10000
)
}
\arguments{
\item{model}{Water model: 1) no water level changes, 2) literature water level change,
3) extrapolated water level changes, 4) using random rates, until literature values.}

\item{crown_age}{age of the crown of the tree}

\item{min_lin}{minimum number of extant lineages in the tree}

\item{max_lin}{maximum number of extant lineages in the tree.}
}
\value{
phy object
}
\description{
simulate a tree using the environmental diversification model
}
