% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_envidiv_tree.R
\name{sim_table_to_newick}
\alias{sim_table_to_newick}
\title{function to transform an Ltable to a newick string. Calls
sim_table_to_phy and then transforms to newick}
\usage{
sim_table_to_newick(input_matrix, crown_age)
}
\arguments{
\item{input_matrix}{input matrix}

\item{crown_age}{crown age}
}
\description{
transform an Ltable to a newick string
}
