% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{smc_abc_batch}
\alias{smc_abc_batch}
\title{simulate a tree using environmental diversification}
\usage{
smc_abc_batch(
  m1,
  m2,
  m3,
  m_weights,
  batch_size,
  crown_age,
  num_lin,
  num_threads,
  sd_p,
  self_prob_m
)
}
\arguments{
\item{m1}{a matrix with particles of model 1}

\item{m2}{a matrix with particles of model 2}

\item{m3}{a matrix with particles of model 3}

\item{m_weights}{a vector with weight of each model}

\item{batch_size}{number of particles to generate}

\item{crown_age}{crown age}

\item{num_lin}{number of lineages to condition on}

\item{num_threads}{number of threads to use}

\item{sd_p}{standard deviation of parameter perturbation}

\item{self_prob_m}{probability of staying in the same model}
}
\value{
list with generated particles
}
\description{
simulate a tree using environmental diversification
}
