% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sq_numbers_cpp_tbb}
\alias{sq_numbers_cpp_tbb}
\title{test multithreaded works!}
\usage{
sq_numbers_cpp_tbb(n, num_threads)
}
\arguments{
\item{n}{size of vector}

\item{num_threads}{number of threads}
}
\value{
vector of squared numbers
}
\description{
test multithreaded works!
}
