% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{square_number}
\alias{square_number}
\title{bogus testing function}
\usage{
square_number(input)
}
\arguments{
\item{input}{double number}
}
\value{
square of input
}
\description{
bogus testing function
}
