% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_envidiv_tbb}
\alias{test_envidiv_tbb}
\title{function to test table conversion}
\usage{
test_envidiv_tbb(model, crown_age)
}
\arguments{
\item{model}{a vector of the four parameters of the model}

\item{crown_age}{crown age of the tree to be simulated}
}
\value{
list
}
\description{
function to test table conversion
}
