% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_tbb}
\alias{test_tbb}
\title{test for tbb implementation}
\usage{
test_tbb(loop_size, num_threads)
}
\arguments{
\item{loop_size}{size of task}

\item{num_threads}{number of threads}
}
\value{
list
}
\description{
test for tbb implementation
}
