% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc_functions.R
\name{within_prior}
\alias{within_prior}
\title{function to determine if the parameter combination is within the prior}
\usage{
within_prior(params)
}
\arguments{
\item{params}{vector of parameters}
}
\value{
true or false
}
\description{
function to determine if the parameter combination is within the prior
}
