% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc_functions.R
\name{calculate_weight}
\alias{calculate_weight}
\title{function to calculate the weight of a particle. The weight depends on the probability of obtaining the specific parameter combination, given all the previous parameters. Hence, it is the sum of the probability for each previous parameter combination to obtain the observed parameter combination.}
\usage{
calculate_weight(params, other_params, weights, sigma)
}
\arguments{
\item{params}{focal parameters for which to estimate the weight}

\item{other_params}{previous params from which this parameter combination is derived}

\item{weights}{previous weights}

\item{sigma}{standard deviation of the perturbation kernel}
}
\value{
weight
}
\description{
function to calculate the weight of a particle. The weight depends on the probability of obtaining the specific parameter combination, given all the previous parameters. Hence, it is the sum of the probability for each previous parameter combination to obtain the observed parameter combination.
}
