% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{create_tree_cpp}
\alias{create_tree_cpp}
\title{simulate a tree using environmental diversification}
\usage{
create_tree_cpp(parameters, waterlevel_changes, seed, crown_age)
}
\arguments{
\item{parameters}{a vector of the four paramaters of the model}

\item{waterlevel_changes}{a vector that indicates the time points of water level changes}

\item{seed}{pseudo-random number generator seed}

\item{crown_age}{crown age of the tree to be simulated}
}
\value{
newick string
}
\description{
simulate a tree using environmental diversification
}
