% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_naive.R
\name{generate_naive}
\alias{generate_naive}
\title{function to perform ABC-SMC}
\usage{
generate_naive(number_of_particles = 1000, min_tips = 50,
  max_tips = 150, model = 1, emp_tree, file_name)
}
\arguments{
\item{number_of_particles}{number of particles used per iteration of
the SMC algorithm}

\item{min_tips}{minimum number of tips}

\item{max_tips}{maximum number of tips}

\item{model}{used water model}

\item{emp_tree}{phy object holding phylogeny of the tree to be fitted on}

\item{file_name}{file name}
}
\value{
a tibble containing the results
}
\description{
function to perform ABC-SMC
}
