% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_stack.R
\name{generate_stack}
\alias{generate_stack}
\title{function to perform ABC-SMC}
\usage{
generate_stack(
  number_of_replicates = 1000,
  parameters = NULL,
  min_tips = 50,
  max_tips = 150,
  emp_tree = NULL,
  crown_age = NULL,
  write_to_file = FALSE,
  file_name = NULL
)
}
\arguments{
\item{number_of_replicates}{number of particles used per iteration of
the SMC algorithm}

\item{parameters}{parameters, a vector with: 1) extinction rate, 2) (sympatric) speciation rate at high water, 3) sympatric speciation rate at low water, 4) allopatric speciation rate at low water, 5) posterior perturbation and 6) the specific model, where 1) model without water level changes, 2) literature water level changes, 3) extrapolated water level changes, 4) standard birth-death model.}

\item{min_tips}{minimum number of tips}

\item{max_tips}{maximum number of tips}

\item{emp_tree}{phy object holding phylogeny of the tree to be fitted on}

\item{crown_age}{crown age}

\item{write_to_file}{boolean, if TRUE, results are written to file.}

\item{file_name}{file name}

\item{model}{used water model}
}
\value{
a tibble containing the results
}
\description{
function to perform ABC-SMC
}
