% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_trees_tbb.R
\name{generate_trees_tbb}
\alias{generate_trees_tbb}
\title{function to perform ABC-SMC}
\usage{
generate_trees_tbb(
  number_of_trees = 1000,
  min_tips = 50,
  max_tips = 150,
  model = NULL,
  crown_age = NULL,
  file_name_trees = "trees.txt",
  file_name_stats = "stats.txt",
  num_threads = -1,
  block_size = 10000
)
}
\arguments{
\item{number_of_trees}{number of trees to generate}

\item{min_tips}{minimum number of tips (inclusive)}

\item{max_tips}{maximum number of tips (inclusive)}

\item{model}{used water model}

\item{crown_age}{crown age}

\item{file_name_trees}{file name to write trees}

\item{file_name_stats}{file name to write stats}

\item{num_threads}{number of threads}

\item{block_size}{maximum number of trees for which summary statistics are
calculated in parallel. Should fit in memory so should not be too big.
Default is 10000}
}
\value{
a matrix with parameter values and the associated summary statistics
}
\description{
function to perform ABC-SMC
}
