% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_trees_tbb.R
\name{generate_trees_tbb}
\alias{generate_trees_tbb}
\title{function to perform ABC-SMC}
\usage{
generate_trees_tbb(
  number_of_trees = 1000,
  min_tips = 50,
  max_tips = 150,
  model = NULL,
  crown_age = NULL,
  file_name = NULL,
  num_threads = -1
)
}
\arguments{
\item{number_of_trees}{number of trees to generate}

\item{min_tips}{minimum number of tips (inclusive)}

\item{max_tips}{maximum number of tips (inclusive)}

\item{model}{used water model}

\item{crown_age}{crown age}

\item{file_name}{file name to write output to}

\item{num_threads}{number of threads}
}
\value{
a matrix with parameter values and the associated summary statistics
}
\description{
function to perform ABC-SMC
}
