% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_trees_tbb.R
\name{generate_trees_tbb}
\alias{generate_trees_tbb}
\title{function to perform ABC-SMC}
\usage{
generate_trees_tbb(
  number_of_trees = 1000,
  min_tips = 50,
  max_tips = 150,
  model = NULL,
  crown_age = NULL,
  write_to_file = TRUE,
  file_name_trees_prefix = "trees",
  file_name_stats = "stats.txt",
  num_threads = -1,
  block_size = 1000,
  foreach_blocksize = 100
)
}
\arguments{
\item{number_of_trees}{number of trees to generate}

\item{min_tips}{minimum number of tips (inclusive)}

\item{max_tips}{maximum number of tips (inclusive)}

\item{model}{used water model}

\item{crown_age}{crown age}

\item{write_to_file}{write output to file, or not. Should always be TRUE,
because when chopping up in blocks, only the last block is returned if FALSE.
Can be used for file-free debugging.}

\item{file_name_trees_prefix}{prefix of file name to write trees to}

\item{file_name_stats}{file name to write stats}

\item{num_threads}{number of threads}

\item{block_size}{maximum number of trees for which summary statistics are
calculated in parallel. Should fit in memory so should not be too big.
Default is 10000}

\item{foreach_blocksize}{size of blocks used in parallel foreach code}

\item{calc_stats}{calculate summary statistics?}
}
\value{
a matrix with parameter values and the associated summary statistics
}
\description{
function to perform ABC-SMC
}
