% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_params.R
\name{infer_params}
\alias{infer_params}
\title{function to perform ABC-SMC}
\usage{
infer_params(number_of_particles, max_iter, sd_params, emp_tree,
  write_to_file = TRUE, seed = NULL)
}
\arguments{
\item{number_of_particles}{number of particles used per iteration of
the SMC algorithm}

\item{max_iter}{maximum number of iterations}

\item{sd_params}{standard deviation of the paramater perturbation kernel}

\item{emp_tree}{phy object holding phylogeny of the tree to be fitted on}

\item{write_to_file}{(boolean) if true, intermediate output is written to file}

\item{seed}{seed of the pseudo random-number generator}
}
\value{
a tibble containing the results
}
\description{
function to perform ABC-SMC
}
