% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_abc_smc.R
\name{perform_abc_smc}
\alias{perform_abc_smc}
\title{perform abc smc}
\usage{
perform_abc_smc(
  emp_tree,
  num_particles,
  num_threads,
  sd_p,
  sd_self,
  min_tips = 40,
  max_tips = 120
)
}
\arguments{
\item{emp_tree}{phylogenetic tree to fit on}

\item{num_particles}{number of particles}

\item{num_threads}{number of threads}

\item{sd_p}{standard deviation of parameter perturbation}

\item{self_prob}{probability of drawing self model}
}
\value{
vector of weights
}
\description{
perform abc smc
}
