% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_reference_table.R
\name{read_reference_table}
\alias{read_reference_table}
\title{collect reference table based on previously generated reference tables for each model}
\usage{
read_reference_table(path)
}
\arguments{
\item{path}{path of directory where each subdirectory [1, 2, 3] is located}
}
\value{
reference table
}
\description{
collect reference table based on previously generated reference tables for each model
}
