% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions.R
\name{sim_envidiv_tree}
\alias{sim_envidiv_tree}
\title{simulate a tree using the environmental diversification model}
\usage{
sim_envidiv_tree(params, crown_age)
}
\arguments{
\item{params}{parameters used to simulate}

\item{crown_age}{age of the crown of the tree}
}
\value{
phy object
}
\description{
simulate a tree using the environmental diversification model
}
