% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_envidiv_tree.R
\name{sim_envidiv_tree}
\alias{sim_envidiv_tree}
\title{simulate a tree using the environmental diversification model}
\usage{
sim_envidiv_tree(params, crown_age, max_lin = 500, abc = FALSE, seed = NULL)
}
\arguments{
\item{params}{parameters used to simulate:
\itemize{
  \item{extinction}{per lineage extinction rate}
  \item{sympatric speciation rate at high water}{per lineage rate of
  speciation when the water level is high}
  \item{sympatric speciation rate at low water}{per lineage rate of
  speciation when the water level is low}
  \item{allopatric speciation rate}{per allopatric pair rate of speciation}
  \item{perturbation}{standard deviation of post-hoc
                      branching time perturbation}
  \item{water model}{Water model: 1) no water level changes, 2) literature
                    water level change, 3) extrapolated water level changes}
}}

\item{crown_age}{age of the crown of the tree}

\item{max_lin}{maximum number of extant lineages in the tree.}

\item{abc}{(boolean) is the tree simulated in an ABC fitting scheme,
or not? additional verbal output is provided if not.}

\item{seed}{random nmber seed}
}
\value{
phy object
}
\description{
simulate a tree using the environmental diversification model
}
